[skip ci]
• pytorchmodelcreator.py:
  move contents to pytorchcomponents.py

• PytorchModelCreator -> PytorchCompositionWrapper
